Reapply changes
