+++ Content.Client/Access/IdCardSystem.cs
@@ -2,6 +2,4 @@
 
 namespace Content.Client.Access;
 
-public sealed class IdCardSystem : SharedIdCardSystem
-{
-}
+public sealed class IdCardSystem : SharedIdCardSystem;
